nuumio: Add RetrogatoR DT
pwm-fan is disabled because it's currently controlled from user space:
RP64 pwm-fan header is not driving a fan directly in RetrogatoR. It's
used as a control signal for a pwm controlled fan instead. A user space
script adjusts pwn period and duty cycle to be suitable. Kernel
controlled pwm would be nice but haven't got that working yet.